+++ akka-http/src/main/scala/akka/http/javadsl/server/ExceptionHandler.scala
@@ -8,7 +8,7 @@ import akka.http.scaladsl.server
 import akka.japi.pf.PFBuilder
 import akka.http.javadsl.settings.RoutingSettings
 import akka.http.impl.util.JavaMapping.Implicits._
-import akka.http.javadsl.RoutingJavaMapping._
+import RoutingJavaMapping._
 
 object ExceptionHandler {
   /**
